+++ test/python/test_on_active_store_path_resolver.py
@@ -1,0 +1,8 @@
+import unittest
+import mock
+
+from wurf.on_active_store_path_resolver import OnActiveStorePathResolver
+
+def test_on_active_store_path_resolver():
+    pass
+    # @todo add tests
