+++ test/python/test_on_passive_load_path_resolver.py
@@ -1,0 +1,8 @@
+import unittest
+import mock
+
+from wurf.on_passive_load_path_resolver import OnPassiveLoadPathResolver
+
+def test_on_passive_load_path_resolver():
+    pass
+    # @todo add tests
